+++ tests/conftest.py
@@ -141,13 +141,25 @@ def attribute() -> Attribute:
 
 
 @pytest.fixture(scope="class")
-def attribute_key() -> str:
+def attribute_key() -> list[str]:
     return ["height"]
 
 
 @pytest.fixture(scope="class")
-def attribute_key_multiple() -> str:
+def attribute_key_multiple() -> list[str]:
     return ["height", "house-number"]
+
+
+@pytest.fixture
+def attribute_filter() -> str:
+    """Custom attribute filter."""
+    return "height=* or building:levels=*"
+
+
+@pytest.fixture
+def attribute_title() -> str:
+    """Attributes title belonging to custom attribute filter (`attribute_filter)`."""
+    return "Height"
 
 
 @pytest.fixture(scope="class")