+++ tests/unittests/api/test_request_models.py
@@ -2,14 +2,10 @@ import pytest
 from pydantic import ValidationError
 
 from ohsome_quality_api.api.request_models import (
-    AttributeCompletenessRequest,
+    AttributeCompletenessFilterRequest,
+    AttributeCompletenessKeyRequest,
     BaseBpolys,
     IndicatorRequest,
-)
-from ohsome_quality_api.utils.exceptions import (
-    GeoJSONError,
-    GeoJSONGeometryTypeError,
-    GeoJSONObjectTypeError,
 )
 
 
@@ -24,22 +20,22 @@ def test_bpolys_valid(
 
 
 def test_bpolys_invalid(feature_collection_invalid):
-    with pytest.raises((GeoJSONError, ValidationError)):
+    with pytest.raises(ValidationError):
         BaseBpolys(bpolys=feature_collection_invalid)
 
 
 def test_bpolys_unsupported_object_type_feature(feature_germany_heidelberg):
-    with pytest.raises((GeoJSONObjectTypeError, ValidationError)):
+    with pytest.raises(ValidationError):
         BaseBpolys(bpolys=feature_germany_heidelberg)
 
 
 def test_bpolys_unsupported_object_type(geojson_unsupported_object_type):
-    with pytest.raises((GeoJSONObjectTypeError, ValidationError)):
+    with pytest.raises(ValidationError):
         BaseBpolys(bpolys=geojson_unsupported_object_type)
 
 
 def test_bpolys_unsupported_geometry_type(feature_collection_unsupported_geometry_type):
-    with pytest.raises((GeoJSONGeometryTypeError, ValidationError)):
+    with pytest.raises(ValidationError):
         BaseBpolys(bpolys=feature_collection_unsupported_geometry_type)
 
 
@@ -58,25 +54,49 @@ def test_indicator_request_invalid_topic(bpolys):
 
 def test_attribute_completeness_missing_attribute(bpolys, topic_key_building_count):
     with pytest.raises(ValueError):
-        AttributeCompletenessRequest(bpolys=bpolys, topic=topic_key_building_count)
+        AttributeCompletenessKeyRequest(bpolys=bpolys, topic=topic_key_building_count)
 
 
 def test_attribute_completeness_invalid_attribute(bpolys, topic_key_building_count):
     with pytest.raises(ValueError):
-        AttributeCompletenessRequest(
+        AttributeCompletenessKeyRequest(
             bpolys=bpolys, topic=topic_key_building_count, attributes="foo"
         )
 
 
-def test_attribute_completeness(bpolys, topic_key_building_count, attribute_key_height):
-    AttributeCompletenessRequest(
-        bpolys=bpolys, topic=topic_key_building_count, attributes=attribute_key_height
+def test_attribute_completeness_single_attribute(
+    bpolys,
+    topic_key_building_count,
+    attribute_key_height,
+):
+    AttributeCompletenessKeyRequest(
+        bpolys=bpolys,
+        topic=topic_key_building_count,
+        attributes=attribute_key_height,
     )
 
 
 def test_attribute_completeness_multiple_attributes(
-    bpolys, topic_key_building_count, attribute_key_multiple
+    bpolys,
+    topic_key_building_count,
+    attribute_key_multiple,
 ):
-    AttributeCompletenessRequest(
-        bpolys=bpolys, topic=topic_key_building_count, attributes=attribute_key_multiple
+    AttributeCompletenessKeyRequest(
+        bpolys=bpolys,
+        topic=topic_key_building_count,
+        attributes=attribute_key_multiple,
+    )
+
+
+def test_attribute_completeness_attribute_filter(
+    bpolys,
+    topic_key_building_count,
+    attribute_filter,
+    attribute_title,
+):
+    AttributeCompletenessFilterRequest(
+        bpolys=bpolys,
+        topic=topic_key_building_count,
+        attribute_filter=attribute_filter,
+        attribute_title=attribute_title,
     )